Fix: pthread exit race condition
Description:
If pthread A is exiting and pthread B is joining A, consider the
situation, A has invoked syscall ExitThread and is recycled. However
B hasn't reached the joining point. Kernel has already created another
thread with the same tid as pthread A and reuse pthread A's PosixTcb.
Thus B would be joining a wrong thread.

Change-Id: Id20ff2403f680cfc85c74dda00ad8810c7734148